+++ extensions/unsure-calc/CHANGELOG.md
@@ -1,3 +1,3 @@
 # Unsure calc Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+## [Initial Version] - 2026-01-08